Fix-A-Lot update
So...sometime on the 15th of February, a ton of aspects of the plugin spontaneously broke. No idea why it happened in the first place, but most of them are now fixed!
- Images & videos that are long and/or have a high width no longer form large, clickable "gaps" below them to fit a square space.
- Videos will properly fill the screen in fullscreen mode again.

These fixes were graciously made by vantage87 on Discord, thanks a ton for their efforts in making up for my cluelessness!

TODOs:  Disabling the video carousel! Hopefully! Maybe!