(mainly) Class Updates
Made a new branch for the version of the plugin with updated classes, as it's still unknown whether Discord plans to revert it. If the class changes become permanent, this branch will be merged with the main one.

Added a piece of code that will work in parity with the UncompressedImages plugin, given it receives patches from its dev. Until then, it's been commented out as it has no use otherwise.